pyproject.toml
[tool.poetry]
name = "dbcat"
version = "0.8.3"
description = "Tokern Data Catalog"
authors = ["Tokern <info@tokern.io>"]
license = "MIT"
classifiers = [
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8",
        "Topic :: Database",
        "Topic :: Software Development",
        "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords=["data-catalog","postgres","snowflake","redshift","glue"]
readme="README.md"
homepage="https://tokern.io/"
repository="https://github.com/tokern/dbcat/"


[tool.poetry.dependencies]
python = "^3.8"
click = "*"
PyYAML = "*"
amundsen-databuilder = {extras = ["athena", "bigquery", "glue", "rds", "snowflake"], version = "^6"}
psycopg2 = "^2.9.1"
alembic = "^1.6.5"
snowflake-sqlalchemy = "1.2.4"
sqlalchemy-mixins = "^1.5"
pyhocon = "^0.3.58"
pyathena = {version = "1.11.5", extras = ["sqlalchemy"]}
PyMySQL = "^1.0.2"
SQLAlchemy = "~=1.3.24"
botocore = "~=1.20.23"
pyparsing = "~=2.0"
boto3 = "1.17.23"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]
black = "==19.10b0"
flake8 = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
types-PyYAML = "^5.4.3"
types-PyMySQL = "^1.0.0"
pytest-cases = "^3.6.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dbcat = "dbcat.__main__:app"
